Adds option to add Tycoon's Wallet to item pool.
It uses what will be the shopsanity cvar, since this will be tied in to shopsanity once that is ready. The idea is that when the full shopsanity GUI is ready and all values are selectable, the Tycoon's Wallet will already be working without needing to change anything about the item pool generation.