Removed indentation spoiler log
Was used to create seperation in the spoiler log, but looked kinda stupid. Also if we want to take these values for titles in the ImGui stuff, it won't look nice.

If we want to do something with this in the spoiler log, it's probably better to format it directly where the spoiler log is created.